upload single-channel yaml file
I uploaded a single-channel version of the yaml file. One of the decisions the group will need to make fairly soon is what information to include in the yaml file and how to organize the data in the file. I've written comments in the file, but essentially should one file have all the channels, some of the channels, or only one channel? In general, we try to calibrate all the channels during one calibration exercise. However, that doesn't always happen and can complicate the mixing of channels that have been calibrated and those that weren't in the same yaml file. 

Anyway, let me know what you think of the file and we can go from there.